feat: かな入力に対応
