+++ Sources/KanaKanjiConverterModule/InputManagement/InputTableID.swift
@@ -3,6 +3,8 @@ public import struct Foundation.URL
 public enum InputTableID: Sendable, Equatable, Hashable {
     case defaultRomanToKana
     case defaultAZIK
+    case defaultKanaJIS
+    case defaultKanaUS
     case empty
     case custom(URL)
 }
